+++ src/langserver/symbolsinformation/toolsGetPermissionSetsListResponse.ts
@@ -1,0 +1,5 @@
+import { PermissionSetInformation } from "../../symbolsinformation/permissionSetInformation";
+import { ToolsSymbolInformationResponse } from "./toolsSymbolInformationResponse";
+
+export class ToolsGetPermissionSetsListResponse extends ToolsSymbolInformationResponse<PermissionSetInformation> {
+}