+++ src/objectwizards/syntaxbuilders/alPermissionSetExtensionSyntaxBuilder.ts
@@ -1,0 +1,24 @@
+import * as vscode from 'vscode';
+import { ALSyntaxWriter } from '../../allanguage/alSyntaxWriter';
+import { ALPermissionSetExtensionWizardData } from "../wizards/alPermissionSetExtensionWizardData";
+import { ALPermissionSetSyntaxBuilder } from "./alPermissionSetSyntaxBuilder";
+
+export class ALPermissionSetExtensionSyntaxBuilder extends ALPermissionSetSyntaxBuilder {
+
+    buildFromPermissionSetExtWizardData(destUri: vscode.Uri | undefined, data : ALPermissionSetExtensionWizardData) : string {
+        let writer : ALSyntaxWriter = new ALSyntaxWriter(destUri);
+       
+        writer.writeStartExtensionObject("permissionsetextension", data.objectId, data.objectName, data.basePermissionSet);
+
+        this.writeIncludePermissionSetList(writer, data.selectedPermissionSetList);
+        this.writePermissions(writer, data.selectedObjectsList);
+
+        writer.writeLine("");
+
+        //finish object
+        writer.writeEndObject();
+        
+        return writer.toString();
+    }
+
+}