+++ src/objectwizards/wizards/alPermissionSetWizardData.ts
@@ -1,0 +1,24 @@
+import { ToolsGetProjectSettingsResponse } from "../../langserver/toolsGetProjectSettingsResponse";
+import { ObjectInformation } from "../../symbolsinformation/objectInformation";
+
+export class ALPermissionSetWizardData {
+    objectId : string;
+    objectName: string;
+    inclAllObjects: boolean;
+
+    projectSettings: ToolsGetProjectSettingsResponse | undefined;
+    permissionSetList : string[] | undefined;
+    selectedPermissionSetList : string[] | undefined;
+    selectedObjectsList: ObjectInformation[] | undefined;
+
+    constructor() {
+        this.objectId = '';
+        this.objectName = '';
+        this.inclAllObjects = true;
+        this.projectSettings = undefined;
+        this.permissionSetList = undefined;
+        this.selectedPermissionSetList = undefined;
+        this.selectedObjectsList = undefined;
+    }
+
+}