+++ src/symbolsinformation/objectInformation.ts
@@ -1,0 +1,5 @@
+import { SymbolWithIdInformation } from "./symbolWithIdInformation";
+
+export class ObjectInformation extends SymbolWithIdInformation {
+    type: string | undefined;
+}